+++ R/foreign-keys.R
@@ -32,19 +32,11 @@
 #' @return For `dm_add_fk()`: An updated `dm` with an additional foreign key relation.
 #'
 #' @export
-#' @examples
-#' if (rlang::is_installed("nycflights13")) {
-#'   nycflights_dm <- dm(
-#'     planes = nycflights13::planes,
-#'     flights = nycflights13::flights
-#'   )
-#' } else {
-#'   message("Using mock-up data, install the nycflights13 package to fix.")
-#'   nycflights_dm <- dm(
-#'     planes = tibble(tailnum = character()),
-#'     flights = tibble(tailnum = character())
-#'   )
-#' }
+#' @examplesIf rlang::is_installed("nycflights13") && rlang::is_installed("DiagrammeR")
+#' nycflights_dm <- dm(
+#'   planes = nycflights13::planes,
+#'   flights = nycflights13::flights
+#' )
 #'
 #' nycflights_dm %>%
 #'   dm_draw()
@@ -116,7 +108,7 @@ dm_add_fk_impl <- function(dm, table, column, ref_table) {
 #' @family foreign key functions
 #'
 #' @export
-#' @examples
+#' @examplesIf rlang::is_installed("nycflights13")
 #' dm_nycflights13() %>%
 #'   dm_has_fk(flights, airports)
 #' dm_nycflights13() %>%
@@ -153,7 +145,7 @@ dm_has_fk_impl <- function(dm, table_name, ref_table_name) {
 #' pointing to the primary key of `ref_table`.
 #'
 #' @export
-#' @examples
+#' @examplesIf rlang::is_installed("nycflights13")
 #' dm_nycflights13() %>%
 #'   dm_get_fk(flights, airports)
 #' dm_nycflights13(cycle = TRUE) %>%
@@ -193,8 +185,9 @@ dm_get_fk_impl <- function(dm, table_name, ref_table_name) {
 #'
 #' @family foreign key functions
 #'
-#' @examples
-#' dm_get_all_fks(dm_nycflights13())
+#' @examplesIf rlang::is_installed("nycflights13")
+#' dm_nycflights13() %>%
+#'   dm_get_all_fks()
 #' @export
 dm_get_all_fks <- function(dm) {
   check_not_zoomed(dm)
@@ -220,8 +213,7 @@ dm_get_all_fks_impl <- function(dm) {
 #' @return For `dm_rm_fk()`: An updated `dm` without the given foreign key relation.
 #'
 #' @export
-#' @examples
-#'
+#' @examplesIf rlang::is_installed("nycflights13") && rlang::is_installed("DiagrammeR")
 #' dm_nycflights13(cycle = TRUE) %>%
 #'   dm_rm_fk(flights, dest, airports) %>%
 #'   dm_draw()
@@ -312,7 +304,7 @@ dm_rm_fk_impl <- function(dm, table_name, cols, ref_table_name) {
 #'
 #' @family foreign key functions
 #'
-#' @examples
+#' @examplesIf rlang::is_installed("nycflights13")
 #' dm_nycflights13() %>%
 #'   dm_enum_fk_candidates(flights, airports)
 #'