+++ R/ptype.R
@@ -12,15 +12,13 @@
 #'
 #' @inheritParams dm_get_fk
 #' @export
-#' @examples
-#' \dontrun{
+#' @examplesIf rlang::is_installed("RMariaDB") && getRversion() >= 3.5
 #' dm_financial() %>%
 #'   dm_ptype()
 #'
 #' dm_financial() %>%
 #'   dm_ptype() %>%
 #'   dm_nrow()
-#' }
 dm_ptype <- function(dm) {
   check_not_zoomed(dm)