+++ tests/testthat/test-zzx-deprecated.R
@@ -19,6 +19,7 @@ test_that("cdm_rm_tbl() works", {
 
 test_that("cdm_copy_to() behaves correctly", {
   skip_on_cran()
+  skip_if_not_installed("dbplyr")
 
   expect_equivalent_dm(
     cdm_copy_to(sqlite(), dm_for_filter(), unique_table_names = TRUE),
@@ -36,6 +37,8 @@ test_that("cdm_disambiguate_cols() works as intended", {
 
 test_that("cdm_get_colors() behaves as intended", {
   skip_on_cran()
+  skip_if_not_installed("nycflights13")
+
   expect_equal(
     cdm_get_colors(cdm_nycflights13()),
     set_names(
@@ -106,6 +109,8 @@ test_that("cdm_get_src() works", {
 })
 
 test_that("cdm_get_con() works", {
+  rlang::local_options(lifecycle_verbosity = "quiet")
+
   skip_on_cran()
 
   expect_dm_error(
@@ -252,10 +257,10 @@ test_that("cdm_paste() works", {
   expect_output(
     cdm_paste(dm_for_filter(), FALSE, 4),
     paste0(
-      "dm(tf_1, tf_2, tf_3, tf_4, tf_5, tf_6) %>%\n    dm_add_pk(tf_1, a) %>%\n    dm_add_pk(tf_2, c) %>%",
-      "\n    dm_add_pk(tf_3, f) %>%\n    dm_add_pk(tf_4, h) %>%\n    dm_add_pk(tf_5, k) %>%\n    ",
-      "dm_add_pk(tf_6, n) %>%\n    dm_add_fk(tf_2, d, tf_1) %>%\n    dm_add_fk(tf_2, e, tf_3) %>%\n    ",
-      "dm_add_fk(tf_4, j, tf_3) %>%\n    dm_add_fk(tf_5, l, tf_4) %>%\n    dm_add_fk(tf_5, m, tf_6)"
+      "dm::dm(tf_1, tf_2, tf_3, tf_4, tf_5, tf_6) %>%\n    dm::dm_add_pk(tf_1, a) %>%\n    dm::dm_add_pk(tf_2, c) %>%",
+      "\n    dm::dm_add_pk(tf_3, f) %>%\n    dm::dm_add_pk(tf_4, h) %>%\n    dm::dm_add_pk(tf_5, k) %>%\n    ",
+      "dm::dm_add_pk(tf_6, n) %>%\n    dm::dm_add_fk(tf_2, d, tf_1) %>%\n    dm::dm_add_fk(tf_2, e, tf_3) %>%\n    ",
+      "dm::dm_add_fk(tf_4, j, tf_3) %>%\n    dm::dm_add_fk(tf_5, l, tf_4) %>%\n    dm::dm_add_fk(tf_5, m, tf_6)"
     ),
     fixed = TRUE
   )